Convert the spawn point type

+++ src/core/world.cpp
@@ -500,7 +500,7 @@ auto world::step() -> void
 
 level::level(std::size_t width, std::size_t height, const std::vector<pixel>& data)
     : pixels{width, height, data}
-    , spawn_point{width / 2, height / 2}
+    , spawn_point{static_cast<int>(width) / 2, static_cast<int>(height) / 2}
     , player{pixels.physics()}
 {
 }
